Add developers section in footer
